+++ index.html
@@ -253,7 +253,6 @@
     </section>
   </template>
 
-  <script src="./js/functions.js"></script>
   <script src="./vendor/nouislider/nouislider.js"></script>
   <script src="./vendor/pristine/pristine.min.js"></script>
   <script src="./js/main.js" type="module"></script>
