+++ js/api.js
@@ -1,24 +1,32 @@
-// const getData = () => fetch(
-//   'https://31.javascript.htmlacademy.pro/kekstagram/data')
-//   .then((response) => {
-//     if (response.ok) {
-//       return response.json();
-//     } else {
-//       throw new Error ('Не удалось загрузить данные');
-//     }
-//   })
-//   .catch(() => {
-//     showDataErrorMessage();
-//   });
+const BASE_URL = 'https://31.javascript.htmlacademy.pro/kekstagram';
+const Route = {
+  GET_DATA: '/data',
+  SEND_DATA: '/',
+};
+const Method = {
+  GET: 'GET',
+  POST: 'POST',
+};
 
+const ErrorText = {
+  GET_DATA: 'Не удалось загрузить данные',
+  SEND_DATA: 'Ошибка загрузки файла',
+};
 
-// const sendData = (body) => fetch (
-//   'https://31.javascript.htmlacademy.pro/kekstagram',
-//   {
-//     method: 'POST',
-//     body,
-//   },
-// );
+const load = (route, errorText, method = Method.GET, body = null) =>
+  fetch(`${BASE_URL}${route}`, {method, body})
+    .then((response) => {
+      if (response.ok) {
+        return response.json();
+      }
+      throw new Error(errorText);
+    })
+    .catch(() => {
+      throw new Error(errorText);
+    });
 
+const getData = () => load(Route.GET_DATA, ErrorText.GET_DATA);
 
-// export { getData, sendData };
+const sendData = (body) => load (Route.SEND_DATA, ErrorText.SEND_DATA, Method.POST, body);
+
+export { getData, sendData };
