+++ js/posted-photo-gallery.js
@@ -1,16 +1,3 @@
-// import { createArrayOfPostedPhotos } from './create-array-of-posted-photos.js';
-
-// let postedPhotos;
-
-// const getPostedPhotos = () => {
-//   if (!postedPhotos) {
-//     postedPhotos = createArrayOfPostedPhotos();
-//   }
-//   return postedPhotos;
-// };
-
-// export { getPostedPhotos };
-
 let postedPhotos = [];
 
 const savePostedPhotos = (loadPhotos) => {
