+++ js/render-thumbnails.js
@@ -1,10 +1,6 @@
-// import { getPostedPhotos } from './posted-photo-gallery.js';
-
 const picturesContainer = document.querySelector('.pictures');
 const templatePictureContent = document.querySelector('#picture').content;
 const templatePicture = templatePictureContent.querySelector('.picture');
-
-// const postedPhotos = getPostedPhotos();
 
 const createThumbnail = ({id, url, description, likes, comments}) => {
   const newThumbnail = templatePicture.cloneNode(true);
@@ -33,18 +29,3 @@ const renderThumbnails = (postedPhotos) => {
 };
 
 export { renderThumbnails };
-
-
-/* Если потребуется универсальная функция для отрисовки элементов
-
-const renderElements = (sourceData, constructorFunction, placeToAdd) => {
-  const fragment = document.createDocumentFragment();
-
-  sourceData.forEach((elem) => {
-    const element = constructorFunction(elem);
-    fragment.append(element);
-  });
-
-  placeToAdd.append(fragment);
-};
-*/
