+++ js/upload-form/change-image-scale.js
@@ -1,5 +1,6 @@
 const imageScale = document.querySelector('.scale__control--value');
 const imagePreview = document.querySelector('.img-upload__preview img');
+const imageUploadScale = document.querySelector('.scale');
 
 const SCALE_STEP = 25;
 const MIN_SCALE = 25;
@@ -33,4 +34,16 @@ const resetImageScale = () => {
   setScaleStyle(DEFAULT_SCALE);
 };
 
-export { doScaleSmaller, doScaleBigger, resetImageScale };
+// Изменение масштаба загружаемой фотографии
+const setImageUploadScale = () => {
+  imageUploadScale.addEventListener('click', (evt) => {
+    if (evt.target.closest('.scale__control--smaller')) {
+      doScaleSmaller();
+    } else if (evt.target.closest('.scale__control--bigger')) {
+      doScaleBigger();
+    }
+  });
+};
+
+
+export { setImageUploadScale, resetImageScale };
