+++ js/upload-form/validate-image-upload-form.js
